requirements.txt
streamlit
easyocr
yake
nltk
google-search-results
Pillow
opencv-python-headless
